+++ unlock-js/src/v02/updateKeyPrice.js
@@ -4,19 +4,19 @@ import TransactionTypes from '../transactionTypes'
 import Errors from '../errors'
 
 /**
- *
- * @param {PropTypes.address} lock : address of the lock for which we update the price
- * @param {PropTypes.address} account: account who owns the lock
- * @param {string} price : new price for the lock
+ * Changes the price of keys on a given lock
+ * @param {object} params
+ * - {PropTypes.address} lockAddress : address of the lock for which we update the price
+ * - {string} price : new price for the lock, as a string (as wei)
  */
-export default async function(lockAddress, account, price) {
+export default async function({ lockAddress, price }) {
   const lockContract = await this.getLockContract(lockAddress)
   let transactionPromise
   try {
     transactionPromise = lockContract['updateKeyPrice(uint256)'](
       utils.toWei(price, 'ether'),
       {
-        gasLimit: GAS_AMOUNTS.updateKeyPrice, // overrides default value for transaction gas price
+        gasLimit: GAS_AMOUNTS.updateKeyPrice,
       }
     )
     const ret = await this._handleMethodCall(