+++ unlock-js/src/v11/withdrawFromLock.js
@@ -5,11 +5,13 @@ import Errors from '../errors'
 
 /**
  * Triggers a transaction to withdraw funds from the lock and assign them to the owner.
- * @param {PropTypes.address} lock
- * @param {PropTypes.string} amount
- * @param {PropTypes.string} decimals
+ * By default (amount=0), this withdraws all funds.
+ * @param {PropTypes.address} lockAddress
+ * @param {string} amount
+ * @param {number} decimals
+ * TODO: get the decimal from the ERC20 contract
  */
-export default async function(lockAddress, amount = '0', decimals = 18) {
+export default async function({ lockAddress, amount = '0', decimals = 18 }) {
   const lockContract = await this.getLockContract(lockAddress)
   let transactionPromise
 
@@ -17,7 +19,7 @@ export default async function(lockAddress, amount = '0', decimals = 18) {
 
   try {
     transactionPromise = lockContract['withdraw(uint256)'](actualAmount, {
-      gasLimit: GAS_AMOUNTS.withdraw, // overrides default value for transaction gas price
+      gasLimit: GAS_AMOUNTS.withdraw,
     })
     const ret = await this._handleMethodCall(
       transactionPromise,