+++ unlock-js/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Version (minor!)
 
 - updateKeyPrice now uses the right decimals for erc20 contracts
+- getting erc20Address and decimals from the contract when purchasing a key
 - Removed the 'owner' param on createLock since it is not really used (just emitted back)
 - Refactored signature to accept objects to be more flexible (this is a breaking change)